requirements.txt
# Runtime dependency for type stubs (development only)
fake-bpy-module-latest

# Development tools
ruff
